+++ epi_judge_python/replace_and_remove.py
@@ -25,33 +25,11 @@ space in the array to hold final result
 - Delete each entry containing b
 
 Hint: consider performing multiple passes on s
-
+[ ] 5/27
 [ ATTEMPTED ] - 5/25
 """
 def replace_and_remove(size, s):
-    write_idx = 0
-    a_count = 0
-    for i in range(size):
-        if s[i] != 'b':
-            s[write_idx] = s[i]
-            write_idx += 1
-        if s[i] == 'a':
-            a_count += 1
-
-    cur_idx = write_idx - 1
-    # the size of the final string
-    write_idx += a_count - 1
-    final_size = write_idx + 1
-    while cur_idx >= 0:
-        if s[cur_idx] == 'a':
-            s[write_idx - 1:write_idx + 1] = "dd"
-            write_idx -= 2
-        else:
-            s[write_idx] = s[cur_idx]
-            write_idx -= 1
-        cur_idx -= 1
-    return final_size
-
+    pass
 
 @enable_executor_hook
 def replace_and_remove_wrapper(executor, size, s):