+++ epi_judge_python_solutions/nearest_repeated_entries.py
@@ -12,6 +12,7 @@ the second and third occurences of "no" is the closest pair
 HISTORY LOG
 - [ SOLVED ] 5/16
 - [ SOLVED ] 5/18
+- [ SOLVED ] 5/27
 """
 def find_nearest_repetition(paragraph):