+++ history.md
@@ -1,6 +1,6 @@
 | Problem |  Attempt Log | Python  |
 | ------ | ------ | ------ |
-| **Bootcamp 6: String** | 5/25 | ------- |
+| **Bootcamp 6: String** | 5/25, 5/27 | ------- |
 | 6.4 Replace and Remove | - *ATTEMPED*, 5/25 | replace_and_remove.py |
 | 6.5 Test palindromicity | - *ATTEMPTED*, 5/25 | is_string_palindromic.py |
 | **Bootcamp 7: Linked List** | 5/25 | ------- |
@@ -8,4 +8,10 @@
 | 7.2 Reverse a single sublist | | reverse\_sublist.py |
 | 7.3 Test for cyclicity | | is\_list\_cyclic.py |
 | 7.4 Test for overlapping lists-lists are cycle-free| | do\_terminated\_lists\_overlap.py |
-| 7.7  Remove kth last element from a list| delete\_kth\_last\_from\_list.py |
+| 7.7  Remove kth last element from a list| delete_kth_last_from_list.py |
+| **Bootcamp 6: Hash Table** | 5/17, 5/27 | ------- |
+| 12.1 Test for palindromic permutations | *SOLVED* 5/17 | is_string_permutable_to_palindrome.py |
+| 12.2 Is an anonymous letter constructible |  *SOLVED* 5/18 | is_anonymous_letter_constructible.py |
+| 12.3 Implement an ISBN cache/LRU Cache |  *SOLVED* 5/18 | lru_cache.py |
+| 12.5 Find the nearest repeated entries in an array | *SOLVED* 5/18 | nearest_repeated_entries.py |
+| 12.6 Find smallest subarray covering all values| *SOLVED* 5/18 | smallest_subarray_covering_all_values.py |